parsing pom.xml
Implement function which parses the pom.xml of a maven project.
It add the java agent command line option ideally to the maven
surefire plugin which is responsible for running unit tests. If the
project does not have a 'argLine' tag it will pass the java agent
argument via the shell command line option 'mvn test -DargLine=...'