Add function doc and return the whole list
The returned list was previously returned without the first entry.
This is due to the assumption that the first line of the csv is
contained in the list but this is wrong. The regular expression already
omits the fist line of the csv which is only the description of
the variables.